item and restaurant services implemented
* added dto objects for restaurant and menu item entites
* implemented restaurant service and controller
* extended item service to also update related restaurant id
* extended item service to get items for a restaurant id

+++ DoyProject/Doy/Doy_Backend/src/main/java/com/pingfloyd/doy/services/IRestaurantService.java
@@ -2,12 +2,10 @@ package com.pingfloyd.doy.services;
 
 import com.pingfloyd.doy.entities.DtoRestaurant;
 import com.pingfloyd.doy.entities.DtoRestaurantIU;
-import com.pingfloyd.doy.entities.Item;
-import org.springframework.web.multipart.MultipartFile;
 
 public interface IRestaurantService {
-    DtoRestaurant getRestaurant(int id);
+    DtoRestaurant getRestaurant(Long id);
     DtoRestaurant postRestaurant(DtoRestaurantIU dtoRestaurantIU);
-    DtoRestaurant updateRestaurant(int id, DtoRestaurantIU dtoRestaurantIU);
-    DtoRestaurant deleteRestaurant(int id);
+    DtoRestaurant updateRestaurant(Long id, DtoRestaurantIU dtoRestaurantIU);
+    DtoRestaurant deleteRestaurant(Long id);
 }
